Profanity Filter Bug fix, Profanity Filter Enabling/Disabling commands, Pony roles (sort of)
I figured it would be best to add Manager/Sweetie Bot technician only
commands to enable and disable to profanity filter. If for no other
reason but to have a quick fix in the case of technical difficulties.

And about the pony roles. Those roles aren't created yet and I'm not
even sure what their final names will be. And I also take issue with how
badly designed the pony role commands are as far as code reusage goes.
It might be better to have a single command that takes in the pony name
as second parameter so that I can use one command for all roles.